requirements.txt
# LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3
langchain-community>=0.0.20

# Vector Store
pinecone-client>=3.0.0

# Document Loaders
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.2

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.0.0
